# swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Traceify GPS API",
    "description": "Interactive testing playground and documentation for the Traceify NestJS + MongoDB GPS tracking backend. Make sure the backend server (default: http://localhost:3001) is running before using the 'Try it out' features.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local NestJS Backend Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/v1/users": {
      "get": {
        "tags": ["Users"],
        "summary": "List users by role",
        "parameters": [
          {
            "name": "role",
            "in": "query",
            "schema": { "type": "string", "default": "customer" }
          }
        ],
        "responses": {
          "200": { "description": "Success" }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Create a new user profile",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["phone_number"],
                "properties": {
                  "phone_number": { "type": "string", "example": "9876543210" },
                  "full_name": { "type": "string", "example": "Chaitra" },
                  "email": { "type": "string", "example": "chaitra@example.com" },
                  "role": { "type": "string", "example": "customer" },
                  "pass_name": { "type": "string", "example": "pass123" },
                  "pass_code": { "type": "string", "example": "code456" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "User created" }
        }
      }
    }
  }
}